Refactor architecture
I thought that DbusBluez storing the BTSensor objects was an awkward
design. The DbusBluez needed to update the BTDevice objects that were
encapsulated inside the BTSensor objects. This commit switches the
BTSensor and BTDevice abstraction layers around, so that DbusBluez now
stores the BTDevice objects that then encapsulate the BTSensor objects.